Refactor planner to prevent thrashing and fix encryption key rotation
