+++ pkg/controllers/provisioningv2/rke2/planner/encryptionkeyrotation.go
@@ -12,7 +12,6 @@ import (
 	rkev1 "github.com/rancher/rancher/pkg/apis/rke.cattle.io/v1"
 	"github.com/rancher/rancher/pkg/apis/rke.cattle.io/v1/plan"
 	"github.com/rancher/rancher/pkg/controllers/provisioningv2/rke2"
-	"github.com/rancher/wrangler/pkg/generic"
 	"github.com/sirupsen/logrus"
 	"k8s.io/apimachinery/pkg/api/equality"
 )
@@ -28,8 +27,6 @@ const (
 
 	encryptionKeyRotationSecretsEncryptApplyCommand  = "secrets-encrypt-apply"
 	encryptionKeyRotationSecretsEncryptStatusCommand = "secrets-encrypt-status"
-
-	encryptionKeyRotationLeaderAnnotation = "rke.cattle.io/encrypt-key-rotation-leader"
 
 	encryptionKeyRotationInstallRoot = "/var/lib/rancher"
 	encryptionKeyRotationBinPrefix   = "rancher_v2prov_encryption_key_rotation/bin"
@@ -126,108 +123,109 @@ echo "$targetGeneration" > "$generationFile"
 	encryptionKeyRotationEndpointEnv = "CONTAINER_RUNTIME_ENDPOINT=unix:///var/run/k3s/containerd/containerd.sock"
 )
 
-func (p *Planner) setEncryptionKeyRotateState(controlPlane *rkev1.RKEControlPlane, status *rkev1.RotateEncryptionKeys, phase rkev1.RotateEncryptionKeysPhase) error {
-	if equality.Semantic.DeepEqual(controlPlane.Status.RotateEncryptionKeys, status) && equality.Semantic.DeepEqual(controlPlane.Status.RotateEncryptionKeysPhase, phase) {
-		return nil
+func (p *Planner) setEncryptionKeyRotateState(status rkev1.RKEControlPlaneStatus, rotate *rkev1.RotateEncryptionKeys, phase rkev1.RotateEncryptionKeysPhase) (rkev1.RKEControlPlaneStatus, error) {
+	if equality.Semantic.DeepEqual(status.RotateEncryptionKeys, status) && equality.Semantic.DeepEqual(status.RotateEncryptionKeysPhase, phase) {
+		return status, nil
 	}
-	controlPlane = controlPlane.DeepCopy()
-	controlPlane.Status.RotateEncryptionKeys = status
-	controlPlane.Status.RotateEncryptionKeysPhase = phase
-	_, err := p.rkeControlPlanes.UpdateStatus(controlPlane)
-	if err != nil {
-		return err
-	}
-	return ErrWaiting("refreshing encryption key rotation state")
+	status.RotateEncryptionKeys = rotate
+	status.RotateEncryptionKeysPhase = phase
+	return status, ErrWaiting("refreshing encryption key rotation state")
 }
 
-func (p *Planner) resetEncryptionKeyRotateState(controlPlane *rkev1.RKEControlPlane) error {
-	if controlPlane.Status.RotateEncryptionKeys == nil && controlPlane.Status.RotateEncryptionKeysPhase == "" {
-		return nil
+func (p *Planner) resetEncryptionKeyRotateState(status rkev1.RKEControlPlaneStatus) (rkev1.RKEControlPlaneStatus, error) {
+	if status.RotateEncryptionKeys == nil && status.RotateEncryptionKeysPhase == "" {
+		return status, nil
 	}
-	return p.setEncryptionKeyRotateState(controlPlane, nil, "")
+	return p.setEncryptionKeyRotateState(status, nil, "")
 }
 
 // rotateEncryptionKeys first verifies that the control plane is in a state where the next step can be derived. If encryption key rotation is required, the corresponding phase and status fields will be set.
 // The function is expected to be called multiple times throughout encryption key rotation, and will set the next corresponding phase based on previous output.
-func (p *Planner) rotateEncryptionKeys(cp *rkev1.RKEControlPlane, releaseData *model.Release, clusterPlan *plan.Plan) error {
+func (p *Planner) rotateEncryptionKeys(cp *rkev1.RKEControlPlane, status rkev1.RKEControlPlaneStatus, releaseData *model.Release, clusterPlan *plan.Plan) (rkev1.RKEControlPlaneStatus, error) {
 	if cp == nil || releaseData == nil || clusterPlan == nil {
-		return fmt.Errorf("cannot pass nil parameters to rotateEncryptionKeys")
+		return status, fmt.Errorf("cannot pass nil parameters to rotateEncryptionKeys")
 	}
 
 	if cp.Spec.RotateEncryptionKeys == nil {
-		return p.resetEncryptionKeyRotateState(cp)
+		return p.resetEncryptionKeyRotateState(status)
 	}
 
 	if supported, err := encryptionKeyRotationSupported(releaseData); err != nil {
-		return err
+		return status, err
 	} else if !supported {
 		logrus.Debugf("rkecluster %s/%s: marking encryption key rotation phase as failed as it was not supported by version: %s", cp.Namespace, cp.Name, cp.Spec.KubernetesVersion)
-		return p.setEncryptionKeyRotateState(cp, cp.Spec.RotateEncryptionKeys, rkev1.RotateEncryptionKeysPhaseFailed)
+		return p.setEncryptionKeyRotateState(status, cp.Spec.RotateEncryptionKeys, rkev1.RotateEncryptionKeysPhaseFailed)
 	}
 
 	if !canRotateEncryptionKeys(cp) {
-		return nil
+		return status, nil
 	}
 
 	if shouldRestartEncryptionKeyRotation(cp) {
 		logrus.Debugf("[planner] rkecluster %s/%s: starting/restarting encryption key rotation", cp.Namespace, cp.Name)
-		return p.setEncryptionKeyRotateState(cp, cp.Spec.RotateEncryptionKeys, rkev1.RotateEncryptionKeysPhasePrepare)
+		return p.setEncryptionKeyRotateState(status, cp.Spec.RotateEncryptionKeys, rkev1.RotateEncryptionKeysPhasePrepare)
 	}
 
-	if err := p.toggleHealthChecksForCluster(cp, true); err != nil {
-		return err
+	if err := p.pauseCAPICluster(cp, true); err != nil {
+		return status, err
 	}
 
-	leader, err := p.encryptionKeyRotationElectLeader(cp, clusterPlan)
+	leader, err := p.encryptionKeyRotationFindLeader(cp, status, clusterPlan)
 	if err != nil {
-		return err
+		return status, err
+	}
+
+	if status.RotateEncryptionKeysLeader != leader.Machine.Name {
+		status.RotateEncryptionKeysLeader = leader.Machine.Name
+		return status, ErrWaitingf("elected %s as control plane leader for encryption key rotation", leader.Machine.Name)
 	}
 
 	logrus.Debugf("[planner] rkecluster %s/%s: current encryption key rotation phase: [%s]", cp.Namespace, cp.Spec.ClusterName, cp.Status.RotateEncryptionKeysPhase)
 
 	switch cp.Status.RotateEncryptionKeysPhase {
 	case rkev1.RotateEncryptionKeysPhasePrepare:
-		err = p.encryptionKeyRotationLeaderPhaseReconcile(cp, leader)
+		status, err = p.encryptionKeyRotationLeaderPhaseReconcile(cp, status, leader)
 		if err != nil {
-			return err
+			return status, err
 		}
-		return p.setEncryptionKeyRotateState(cp, cp.Spec.RotateEncryptionKeys, rkev1.RotateEncryptionKeysPhasePostPrepareRestart)
+		return p.setEncryptionKeyRotateState(status, cp.Spec.RotateEncryptionKeys, rkev1.RotateEncryptionKeysPhasePostPrepareRestart)
 	case rkev1.RotateEncryptionKeysPhasePostPrepareRestart:
-		err = p.encryptionKeyRotationRestartNodes(cp, clusterPlan, leader)
+		status, err = p.encryptionKeyRotationRestartNodes(cp, status, clusterPlan, leader)
 		if err != nil {
-			return err
+			return status, err
 		}
-		return p.setEncryptionKeyRotateState(cp, cp.Spec.RotateEncryptionKeys, rkev1.RotateEncryptionKeysPhaseRotate)
+		return p.setEncryptionKeyRotateState(status, cp.Spec.RotateEncryptionKeys, rkev1.RotateEncryptionKeysPhaseRotate)
 	case rkev1.RotateEncryptionKeysPhaseRotate:
-		err = p.encryptionKeyRotationLeaderPhaseReconcile(cp, leader)
+		status, err = p.encryptionKeyRotationLeaderPhaseReconcile(cp, status, leader)
 		if err != nil {
-			return err
+			return status, err
 		}
-		return p.setEncryptionKeyRotateState(cp, cp.Spec.RotateEncryptionKeys, rkev1.RotateEncryptionKeysPhasePostRotateRestart)
+		return p.setEncryptionKeyRotateState(status, cp.Spec.RotateEncryptionKeys, rkev1.RotateEncryptionKeysPhasePostRotateRestart)
 	case rkev1.RotateEncryptionKeysPhasePostRotateRestart:
-		err = p.encryptionKeyRotationRestartNodes(cp, clusterPlan, leader)
+		status, err = p.encryptionKeyRotationRestartNodes(cp, status, clusterPlan, leader)
 		if err != nil {
-			return err
+			return status, err
 		}
-		return p.setEncryptionKeyRotateState(cp, cp.Spec.RotateEncryptionKeys, rkev1.RotateEncryptionKeysPhaseReencrypt)
+		return p.setEncryptionKeyRotateState(status, cp.Spec.RotateEncryptionKeys, rkev1.RotateEncryptionKeysPhaseReencrypt)
 	case rkev1.RotateEncryptionKeysPhaseReencrypt:
-		err = p.encryptionKeyRotationLeaderPhaseReconcile(cp, leader)
+		status, err = p.encryptionKeyRotationLeaderPhaseReconcile(cp, status, leader)
 		if err != nil {
-			return err
+			return status, err
 		}
-		return p.setEncryptionKeyRotateState(cp, cp.Spec.RotateEncryptionKeys, rkev1.RotateEncryptionKeysPhasePostReencryptRestart)
+		return p.setEncryptionKeyRotateState(status, cp.Spec.RotateEncryptionKeys, rkev1.RotateEncryptionKeysPhasePostReencryptRestart)
 	case rkev1.RotateEncryptionKeysPhasePostReencryptRestart:
-		err = p.encryptionKeyRotationRestartNodes(cp, clusterPlan, leader)
+		status, err = p.encryptionKeyRotationRestartNodes(cp, status, clusterPlan, leader)
 		if err != nil {
-			return err
+			return status, err
 		}
-		if err = p.toggleHealthChecksForCluster(cp, true); err != nil {
-			return err
+		if err = p.pauseCAPICluster(cp, false); err != nil {
+			return status, err
 		}
-		return p.setEncryptionKeyRotateState(cp, cp.Spec.RotateEncryptionKeys, rkev1.RotateEncryptionKeysPhaseDone)
+		status.RotateEncryptionKeysLeader = ""
+		return p.setEncryptionKeyRotateState(status, cp.Spec.RotateEncryptionKeys, rkev1.RotateEncryptionKeysPhaseDone)
 	}
 
-	return fmt.Errorf("encountered unknown encryption key rotation phase: %s", cp.Status.RotateEncryptionKeysPhase)
+	return status, fmt.Errorf("encountered unknown encryption key rotation phase: %s", cp.Status.RotateEncryptionKeysPhase)
 }
 
 // encryptionKeyRotationSupported returns a boolean indicating whether encryption key rotation is supported by the release,
@@ -302,50 +300,43 @@ func rotateEncryptionKeyInProgress(cp *rkev1.RKEControlPlane) bool {
 		cp.Status.RotateEncryptionKeysPhase == rkev1.RotateEncryptionKeysPhasePostReencryptRestart
 }
 
-// encryptionKeyRotationElectLeader returns the current encryption rotation leader if it is valid, otherwise, if the
+// encryptionKeyRotationFindLeader returns the current encryption rotation leader if it is valid, otherwise, if the
 // phase is "prepare", it will re-elect a new leader. It will look for the init node, and if the init node is not valid
 // (etcd-only), it will elect the first suitable control plane node. If the phase is not in "prepare" and a re-election
 // of the leader is necessary, the phase will be set to failed as this is unexpected.
-func (p *Planner) encryptionKeyRotationElectLeader(cp *rkev1.RKEControlPlane, clusterPlan *plan.Plan) (*planEntry, error) {
-	_, _, init, err := p.findInitNode(cp, clusterPlan)
+func (p *Planner) encryptionKeyRotationFindLeader(cp *rkev1.RKEControlPlane, status rkev1.RKEControlPlaneStatus, clusterPlan *plan.Plan) (*planEntry, error) {
+	_, _, init, _ := p.findInitNode(cp, clusterPlan)
 	if init == nil {
-		return nil, p.encryptionKeyRotationFailed(cp, err)
+		return nil, fmt.Errorf("cannot find suitable init node")
 	}
 
-	if machineName, ok := cp.Annotations[encryptionKeyRotationLeaderAnnotation]; ok {
-		if machine, ok := clusterPlan.Machines[machineName]; ok {
-			entry := &planEntry{
-				Machine:  machine,
-				Plan:     clusterPlan.Nodes[machineName],
-				Metadata: clusterPlan.Metadata[machineName],
-			}
-			if encryptionKeyRotationIsSuitableControlPlane(entry) {
-				return entry, nil
-			}
+	machineName := status.RotateEncryptionKeysLeader
+	if machine, ok := clusterPlan.Machines[machineName]; ok {
+		entry := &planEntry{
+			Machine:  machine,
+			Plan:     clusterPlan.Nodes[machineName],
+			Metadata: clusterPlan.Metadata[machineName],
+		}
+		if encryptionKeyRotationIsSuitableControlPlane(entry) {
+			return entry, nil
 		}
 	}
 
-	if cp.Status.RotateEncryptionKeysPhase != rkev1.RotateEncryptionKeysPhasePrepare {
+	if status.RotateEncryptionKeysPhase != rkev1.RotateEncryptionKeysPhasePrepare {
 		// if we are electing a leader and are not in the "prepare" phase, something is wrong.
-		return nil, p.encryptionKeyRotationFailed(cp, fmt.Errorf("re-election of encryption key rotation leader occurred when phase was %s", cp.Status.RotateEncryptionKeysPhase))
+		return nil, fmt.Errorf("cannot elect control plane leader in phase %s", status.RotateEncryptionKeysPhase)
 	}
 
 	leader := init
 	if !isControlPlane(init) {
 		machines := collect(clusterPlan, encryptionKeyRotationIsSuitableControlPlane)
 		if len(machines) == 0 {
-			return nil, p.encryptionKeyRotationFailed(cp, fmt.Errorf("no suitable control plane nodes for encryption key rotation"))
+			return nil, fmt.Errorf("no suitable control plane nodes for encryption key rotation")
 		}
 		leader = machines[0]
 	}
 
-	cp = cp.DeepCopy()
-	cp.Annotations[encryptionKeyRotationLeaderAnnotation] = leader.Machine.Name
-	_, err = p.rkeControlPlanes.Update(cp)
-	if err != nil {
-		return nil, err
-	}
-	return nil, generic.ErrSkip
+	return leader, nil
 }
 
 // encryptionKeyRotationIsSuitableControlPlane ensures that a control plane node has not been deleted and has a valid
@@ -358,7 +349,7 @@ func encryptionKeyRotationIsSuitableControlPlane(entry *planEntry) bool {
 func encryptionKeyRotationIsControlPlaneAndNotLeaderAndInit(cp *rkev1.RKEControlPlane) roleFilter {
 	return func(entry *planEntry) bool {
 		return isControlPlaneAndNotInitNode(entry) &&
-			cp.Annotations[encryptionKeyRotationLeaderAnnotation] != entry.Machine.Name
+			cp.Status.RotateEncryptionKeysLeader != entry.Machine.Name
 	}
 }
 
@@ -366,7 +357,7 @@ func encryptionKeyRotationIsControlPlaneAndNotLeaderAndInit(cp *rkev1.RKEControl
 func encryptionKeyRotationIsEtcdAndNotControlPlaneAndNotLeaderAndInit(cp *rkev1.RKEControlPlane) roleFilter {
 	return func(entry *planEntry) bool {
 		return isEtcd(entry) && !isControlPlane(entry) &&
-			cp.Annotations[encryptionKeyRotationLeaderAnnotation] != entry.Machine.Name &&
+			cp.Status.RotateEncryptionKeysLeader != entry.Machine.Name &&
 			!isInitNode(entry)
 	}
 }
@@ -375,58 +366,59 @@ func encryptionKeyRotationIsEtcdAndNotControlPlaneAndNotLeaderAndInit(cp *rkev1.
 // The followers (if any exist) are subsequently restarted. Notably, if the encryption key rotation leader is not the init node,
 // it will restart the init node, then restart the encryption key rotation leader,
 // then finalize walking through etcd nodes (that are not controlplane), then finally controlplane nodes.
-func (p *Planner) encryptionKeyRotationRestartNodes(cp *rkev1.RKEControlPlane, clusterPlan *plan.Plan, leader *planEntry) error {
+func (p *Planner) encryptionKeyRotationRestartNodes(cp *rkev1.RKEControlPlane, status rkev1.RKEControlPlaneStatus, clusterPlan *plan.Plan, leader *planEntry) (rkev1.RKEControlPlaneStatus, error) {
 	// in certain cases with multi-node setups, we must restart the init node before we can proceed to restarting the leader.
 	if !isInitNode(leader) {
-		logrus.Debugf("[planner] rkecluster %s/%s: leader %s was not the init node, finding and restarting services on init node", cp.Namespace, cp.Name, leader.Machine.Name)
+		logrus.Debugf("[planner] rkecluster %s/%s: leader %s was not the init node, finding and restarting etcd nodes", cp.Namespace, cp.Name, leader.Machine.Name)
 		// find init node and restart service on the init node first.
 		initNodeFound, _, initNode, err := p.findInitNode(cp, clusterPlan)
 		if err != nil {
-			return err
+			return status, err
 		}
-		if !initNodeFound {
-			return fmt.Errorf("unable to find init node")
+		if !initNodeFound || initNode == nil {
+			return status, fmt.Errorf("unable to find init node")
 		}
 
-		_, err = p.encryptionKeyRotationRestartService(cp, initNode, false, "")
+		_, status, err = p.encryptionKeyRotationRestartService(cp, status, initNode, false, "")
 		if err != nil {
-			return err
+			return status, err
 		}
 		logrus.Debugf("[planner] rkecluster %s/%s: collecting etcd and not control plane", cp.Namespace, cp.Name)
 		for _, entry := range collect(clusterPlan, encryptionKeyRotationIsEtcdAndNotControlPlaneAndNotLeaderAndInit(cp)) {
-			_, err = p.encryptionKeyRotationRestartService(cp, entry, false, "")
+			_, status, err = p.encryptionKeyRotationRestartService(cp, status, entry, false, "")
 			if err != nil {
-				return err
+				return status, err
 			}
 		}
 	}
 
-	leaderStage, err := p.encryptionKeyRotationRestartService(cp, leader, true, "")
+	leaderStage, status, err := p.encryptionKeyRotationRestartService(cp, status, leader, true, "")
 	if err != nil {
-		return err
+		return status, err
 	}
 
 	logrus.Debugf("[planner] rkecluster %s/%s: collecting control plane and not leader and init nodes", cp.Namespace, cp.Name)
 	for _, entry := range collect(clusterPlan, encryptionKeyRotationIsControlPlaneAndNotLeaderAndInit(cp)) {
-		stage, err := p.encryptionKeyRotationRestartService(cp, entry, true, leaderStage)
+		var stage string
+		stage, status, err = p.encryptionKeyRotationRestartService(cp, status, entry, true, leaderStage)
 		if err != nil {
-			return err
+			return status, err
 		}
 
 		if stage != leaderStage {
 			// secrets-encrypt command was run on another node. this is considered a failure, but might be a bit too sensitive. to be tested.
-			return p.encryptionKeyRotationFailed(cp, fmt.Errorf("leader [%s] with %s stage and follower [%s] with %s stage", leader.Machine.Status.NodeRef.Name, leaderStage, entry.Machine.Status.NodeRef.Name, stage))
+			return p.encryptionKeyRotationFailed(status, fmt.Errorf("leader [%s] with %s stage and follower [%s] with %s stage", leader.Machine.Status.NodeRef.Name, leaderStage, entry.Machine.Status.NodeRef.Name, stage))
 		}
 	}
 
-	return nil
+	return status, nil
 }
 
 // encryptionKeyRotationRestartService restarts the server unit on the downstream node, waits until secrets-encrypt
 // status can be successfully queried, and then gets the status. leaderStage is allowed to be empty if entry is the
 // leader.
-func (p *Planner) encryptionKeyRotationRestartService(cp *rkev1.RKEControlPlane, entry *planEntry, scrapeStage bool, leaderStage string) (string, error) {
-	runtime := rke2.GetRuntime(cp.Spec.KubernetesVersion)
+func (p *Planner) encryptionKeyRotationRestartService(cp *rkev1.RKEControlPlane, status rkev1.RKEControlPlaneStatus, entry *planEntry, scrapeStage bool, leaderStage string) (string, rkev1.RKEControlPlaneStatus, error) {
+	//runtime := rke2.GetRuntime(cp.Spec.KubernetesVersion)
 	nodePlan := plan.NodePlan{
 		Files: []plan.File{
 			{
@@ -440,7 +432,7 @@ func (p *Planner) encryptionKeyRotationRestartService(cp *rkev1.RKEControlPlane,
 		},
 		Instructions: []plan.OneTimeInstruction{
 			encryptionKeyRotationWaitForSystemctlStatusInstruction(cp),
-			generateKillAllInstruction(runtime),
+			//generateKillAllInstruction(runtime),
 			encryptionKeyRotationRestartInstruction(cp),
 		},
 	}
@@ -466,29 +458,30 @@ func (p *Planner) encryptionKeyRotationRestartService(cp *rkev1.RKEControlPlane,
 	err := assignAndCheckPlan(p.store, fmt.Sprintf("encryption key rotation [%s] for machine [%s]", cp.Status.RotateEncryptionKeysPhase, entry.Machine.Name), entry, nodePlan, 5, 5)
 	if err != nil {
 		if isErrWaiting(err) {
-			return "", err
+			return "", status, err
 		}
-		return "", p.encryptionKeyRotationFailed(cp, err)
+		status, err := p.encryptionKeyRotationFailed(status, err)
+		return "", status, err
 	}
 
 	if !scrapeStage || !isControlPlane(entry) {
-		return "", nil
+		return "", status, nil
 	}
 
 	stage, err := encryptionKeyRotationSecretsEncryptStageFromOneTimeStatus(entry)
 	if err != nil {
-		return "", p.enqueueIfErrWaiting(cp, err)
+		return "", status, err
 	}
-	return stage, nil
+	return stage, status, nil
 }
 
 // encryptionKeyRotationLeaderPhaseReconcile will run the secrets-encrypt command that corresponds to the phase, and scrape output to ensure that it was
 // successful. If the secrets-encrypt command does not exist on the plan, that means this is the first reconciliation, and
 // it must be added, otherwise reenqueue until the plan is in sync.
-func (p *Planner) encryptionKeyRotationLeaderPhaseReconcile(cp *rkev1.RKEControlPlane, leader *planEntry) error {
+func (p *Planner) encryptionKeyRotationLeaderPhaseReconcile(cp *rkev1.RKEControlPlane, status rkev1.RKEControlPlaneStatus, leader *planEntry) (rkev1.RKEControlPlaneStatus, error) {
 	apply, err := encryptionKeyRotationSecretsEncryptInstruction(cp)
 	if err != nil {
-		return p.encryptionKeyRotationFailed(cp, err)
+		return p.encryptionKeyRotationFailed(status, err)
 	}
 
 	nodePlan := plan.NodePlan{
@@ -521,44 +514,30 @@ func (p *Planner) encryptionKeyRotationLeaderPhaseReconcile(cp *rkev1.RKEControl
 			if strings.HasPrefix(err.Error(), "starting") {
 				logrus.Infof("[planner] rkecluster %s/%s: applying encryption key rotation stage command: [%s]", cp.Namespace, cp.Spec.ClusterName, apply.Args[1])
 			}
-			return err
+			return status, err
 		}
-		return p.encryptionKeyRotationFailed(cp, err)
+		return p.encryptionKeyRotationFailed(status, err)
 	}
 	scrapedStageFromOneTimeInstructions, err := encryptionKeyRotationSecretsEncryptStageFromOneTimeStatus(leader)
 	if err != nil {
-		return err
+		return status, err
 	}
 	periodic, err := encryptionKeyRotationSecretsEncryptStageFromPeriodic(leader)
 	if err != nil {
-		return err
+		return status, err
 	}
 	err = encryptionKeyRotationIsCurrentStageAllowed(scrapedStageFromOneTimeInstructions, cp.Status.RotateEncryptionKeysPhase)
 	if err != nil {
-		return p.encryptionKeyRotationFailed(cp, err)
+		return p.encryptionKeyRotationFailed(status, err)
 	}
 	if scrapedStageFromOneTimeInstructions == encryptionKeyRotationStageReencryptRequest || scrapedStageFromOneTimeInstructions == encryptionKeyRotationStageReencryptActive {
 		if periodic != encryptionKeyRotationStageReencryptFinished {
-			return ErrWaitingf("waiting for encryption key rotation stage to be finished")
+			return status, ErrWaitingf("waiting for encryption key rotation stage to be finished")
 		}
 	}
 	// successful restart, complete same phases for rotate & reencrypt
 	logrus.Infof("[planner] rkecluster %s/%s: successfully applied encryption key rotation stage command: [%s]", cp.Namespace, cp.Spec.ClusterName, leader.Plan.Plan.Instructions[0].Args[1])
-	return nil
-}
-
-// encryptionKeyRotationUpdateControlPlanePhase updates the control plane status if it has not been done yet.
-func (p *Planner) encryptionKeyRotationUpdateControlPlanePhase(cp *rkev1.RKEControlPlane, phase rkev1.RotateEncryptionKeysPhase) error {
-	if cp.Status.RotateEncryptionKeysPhase != phase {
-		cp = cp.DeepCopy()
-		logrus.Infof("[planner] rkecluster %s/%s: applying encryption key rotation phase: [%s]", cp.Namespace, cp.Spec.ClusterName, phase)
-		cp.Status.RotateEncryptionKeysPhase = phase
-		_, err := p.rkeControlPlanes.UpdateStatus(cp)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
+	return status, nil
 }
 
 // encryptionKeyRotationSecretsEncryptStageFromPeriodic will attempt to extract the current stage (secrets-encrypt status) from the
@@ -791,14 +770,9 @@ func encryptionKeyRotationIsCurrentStageAllowed(leaderStage string, currentPhase
 
 // encryptionKeyRotationFailed updates the various status objects on the control plane, allowing the cluster to
 // continue the reconciliation loop. Encryption key rotation will not be restarted again until requested.
-func (p *Planner) encryptionKeyRotationFailed(cp *rkev1.RKEControlPlane, err error) error {
-	if err2 := p.setEncryptionKeyRotateState(cp, cp.Spec.RotateEncryptionKeys, rkev1.RotateEncryptionKeysPhaseFailed); err2 != nil {
-		return err2
-	}
-
-	err = errors.Wrap(err, "encryption key rotation failed, please perform an etcd restore")
-	logrus.Errorf("[planner] rkecluster %s/%s: %v", cp.Namespace, cp.Spec.ClusterName, err)
-	return err
+func (p *Planner) encryptionKeyRotationFailed(status rkev1.RKEControlPlaneStatus, err error) (rkev1.RKEControlPlaneStatus, error) {
+	status.RotateEncryptionKeysPhase = rkev1.RotateEncryptionKeysPhaseFailed
+	return status, errors.Wrap(err, "encryption key rotation failed, please perform an etcd restore")
 }
 
 func encryptionKeyRotationScriptPath(cp *rkev1.RKEControlPlane, file string) string {
