+++ pkg/controllers/provisioningv2/rke2/planner/etcdcreate.go
@@ -7,32 +7,31 @@ import (
 	rkev1 "github.com/rancher/rancher/pkg/apis/rke.cattle.io/v1"
 	"github.com/rancher/rancher/pkg/apis/rke.cattle.io/v1/plan"
 	"github.com/rancher/rancher/pkg/controllers/provisioningv2/rke2"
+	"github.com/rancher/wrangler/pkg/merr"
 	"k8s.io/apimachinery/pkg/api/equality"
 )
 
-func (p *Planner) setEtcdSnapshotCreateState(controlPlane *rkev1.RKEControlPlane, spec *rkev1.ETCDSnapshotCreate, phase rkev1.ETCDSnapshotPhase) error {
-	controlPlane = controlPlane.DeepCopy()
-	controlPlane.Status.ETCDSnapshotCreatePhase = phase
-	controlPlane.Status.ETCDSnapshotCreate = spec
-	_, err := p.rkeControlPlanes.UpdateStatus(controlPlane)
-	if err != nil {
-		return err
+func (p *Planner) setEtcdSnapshotCreateState(status rkev1.RKEControlPlaneStatus, create *rkev1.ETCDSnapshotCreate, phase rkev1.ETCDSnapshotPhase) (rkev1.RKEControlPlaneStatus, error) {
+	if status.ETCDSnapshotCreatePhase != phase || !equality.Semantic.DeepEqual(status.ETCDSnapshotCreate, create) {
+		status.ETCDSnapshotCreatePhase = phase
+		status.ETCDSnapshotCreate = create
+		return status, ErrWaiting("refreshing etcd create state")
 	}
-	return ErrWaiting("refreshing etcd create state")
+	return status, nil
 }
 
-func (p *Planner) resetEtcdSnapshotCreateState(controlPlane *rkev1.RKEControlPlane) error {
-	if controlPlane.Status.ETCDSnapshotCreate == nil && controlPlane.Status.ETCDSnapshotCreatePhase == "" {
-		return nil
+func (p *Planner) resetEtcdSnapshotCreateState(status rkev1.RKEControlPlaneStatus) (rkev1.RKEControlPlaneStatus, error) {
+	if status.ETCDSnapshotCreate == nil && status.ETCDSnapshotCreatePhase == "" {
+		return status, nil
 	}
-	return p.setEtcdSnapshotCreateState(controlPlane, nil, "")
+	return p.setEtcdSnapshotCreateState(status, nil, "")
 }
 
-func (p *Planner) startOrRestartEtcdSnapshotCreate(controlPlane *rkev1.RKEControlPlane, snapshot *rkev1.ETCDSnapshotCreate) error {
-	if controlPlane.Status.ETCDSnapshotCreate == nil || !equality.Semantic.DeepEqual(*snapshot, *controlPlane.Status.ETCDSnapshotCreate) {
-		return p.setEtcdSnapshotCreateState(controlPlane, snapshot, rkev1.ETCDSnapshotPhaseStarted)
+func (p *Planner) startOrRestartEtcdSnapshotCreate(status rkev1.RKEControlPlaneStatus, snapshot *rkev1.ETCDSnapshotCreate) (rkev1.RKEControlPlaneStatus, error) {
+	if status.ETCDSnapshotCreate == nil || !equality.Semantic.DeepEqual(*snapshot, *status.ETCDSnapshotCreate) {
+		return p.setEtcdSnapshotCreateState(status, snapshot, rkev1.ETCDSnapshotPhaseStarted)
 	}
-	return nil
+	return status, nil
 }
 
 func (p *Planner) runEtcdSnapshotCreate(controlPlane *rkev1.RKEControlPlane, clusterPlan *plan.Plan) []error {
@@ -52,7 +51,7 @@ func (p *Planner) runEtcdSnapshotCreate(controlPlane *rkev1.RKEControlPlane, clu
 		if server.Machine.Status.NodeRef != nil && server.Machine.Status.NodeRef.Name != "" {
 			msg = fmt.Sprintf("etcd snapshot on node %s", server.Machine.Status.NodeRef.Name)
 		}
-		if err := assignAndCheckPlan(p.store, msg, server, createPlan, 3, 3); err != nil {
+		if err = assignAndCheckPlan(p.store, msg, server, createPlan, 3, 3); err != nil {
 			errs = append(errs, err)
 		}
 	}
@@ -75,18 +74,17 @@ func (p *Planner) generateEtcdSnapshotCreatePlan(controlPlane *rkev1.RKEControlP
 	})
 }
 
-func (p *Planner) createEtcdSnapshot(controlPlane *rkev1.RKEControlPlane, tokensSecret plan.Secret, clusterPlan *plan.Plan) []error {
+func (p *Planner) createEtcdSnapshot(controlPlane *rkev1.RKEControlPlane, status rkev1.RKEControlPlaneStatus, tokensSecret plan.Secret, clusterPlan *plan.Plan) (rkev1.RKEControlPlaneStatus, error) {
+	var err error
 	if controlPlane.Spec.ETCDSnapshotCreate == nil {
-		if err := p.resetEtcdSnapshotCreateState(controlPlane); err != nil {
-			return []error{err}
-		}
-		return nil
+		status, err := p.resetEtcdSnapshotCreateState(status)
+		return status, err
 	}
 
 	snapshot := controlPlane.Spec.ETCDSnapshotCreate
 
-	if err := p.startOrRestartEtcdSnapshotCreate(controlPlane, snapshot); err != nil {
-		return []error{err}
+	if status, err = p.startOrRestartEtcdSnapshotCreate(status, snapshot); err != nil {
+		return status, err
 	}
 
 	switch controlPlane.Status.ETCDSnapshotCreatePhase {
@@ -99,11 +97,11 @@ func (p *Planner) createEtcdSnapshot(controlPlane *rkev1.RKEControlPlane, tokens
 					continue
 				}
 				finErrs = append(finErrs, err)
-				var errWaiting ErrWaiting
-				if !errors.As(err, &errWaiting) {
+				if isErrWaiting(err) {
 					// we have a failed snapshot from a node.
 					if !stateSet {
-						if err := p.setEtcdSnapshotCreateState(controlPlane, snapshot, rkev1.ETCDSnapshotPhaseFailed); err != nil {
+						status, err = p.setEtcdSnapshotCreateState(status, snapshot, rkev1.ETCDSnapshotPhaseFailed)
+						if err != nil {
 							finErrs = append(finErrs, err)
 						} else {
 							stateSet = true
@@ -111,28 +109,28 @@ func (p *Planner) createEtcdSnapshot(controlPlane *rkev1.RKEControlPlane, tokens
 					}
 				}
 			}
-			return finErrs
+			return status, ErrWaiting(merr.NewErrors(finErrs...).Error())
 		}
-		if err := p.setEtcdSnapshotCreateState(controlPlane, snapshot, rkev1.ETCDSnapshotPhaseRestartCluster); err != nil {
-			return []error{err}
+		if status, err = p.setEtcdSnapshotCreateState(status, snapshot, rkev1.ETCDSnapshotPhaseRestartCluster); err != nil {
+			return status, err
 		}
-		return nil
+		return status, nil
 	case rkev1.ETCDSnapshotPhaseRestartCluster:
-		if err := p.runEtcdSnapshotManagementServiceStart(controlPlane, tokensSecret, clusterPlan, isEtcd, "etcd snapshot creation"); err != nil {
-			return []error{err}
+		if err = p.runEtcdSnapshotManagementServiceStart(controlPlane, tokensSecret, clusterPlan, isEtcd, "etcd snapshot creation"); err != nil {
+			return status, err
 		}
-		if err := p.setEtcdSnapshotCreateState(controlPlane, snapshot, rkev1.ETCDSnapshotPhaseFinished); err != nil {
-			return []error{err}
+		if status, err = p.setEtcdSnapshotCreateState(status, snapshot, rkev1.ETCDSnapshotPhaseFinished); err != nil {
+			return status, err
 		}
-		return nil
+		return status, nil
 	case rkev1.ETCDSnapshotPhaseFailed:
 		fallthrough
 	case rkev1.ETCDSnapshotPhaseFinished:
-		return nil
+		return status, nil
 	default:
-		if err := p.setEtcdSnapshotCreateState(controlPlane, snapshot, rkev1.ETCDSnapshotPhaseStarted); err != nil {
-			return []error{err}
+		if status, err = p.setEtcdSnapshotCreateState(status, snapshot, rkev1.ETCDSnapshotPhaseStarted); err != nil {
+			return status, err
 		}
-		return nil
+		return status, nil
 	}
 }
